feat: separate models and client folders
