+++ crates/core/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod client;
 pub mod mock;
+pub mod models;
